PyF

Description

A lot of quasiquoters to format and interpolate string expression

Documentation

fmt :: QuasiQuoter

Generic formatter, can format an expression to any t as long as t is an instance of IsString.

fmtTrim :: QuasiQuoter

Format with whitespace trimming.

str :: QuasiQuoter

multiline string, no interpolation.

strTrim :: QuasiQuoter

multiline string, no interpolation, but does indentation trimming.

raw :: QuasiQuoter

Raw string, no interpolation neither escaping is performed.

module PyF.Class

Whitespace utilities

trimIndent :: String -> String

Removes the trailing whitespace of a string.

First line is ignored if it only contains whitespaces

All other line common indentation is removed, ignoring line with only whitespaces.

>>> trimIndent "\n   hello\n   - a\n   - b\n   "
"hello\n- a\n- b\n"

See fmtTrim for a quasiquoter with this behavior

Configuration

mkFormatter :: String -> Config -> QuasiQuoter

Build a formatter. See the Config type for details, as well as fmtConfig and strConfig for examples.

defaultConfig :: Config

This is an empty configuration. No formatting, no post processing

fmtConfig :: Config

The config for fmt.

strConfig :: Config

Configuration for str it just wrap the multiline string with fromString.

addTrim :: Config -> Config

Add indentation trimming to a configuration.

addFormatting :: (Char, Char) -> Config -> Config

Enable formatting