PyF.Class

Description

You want to add formatting support for your custom type. This is the right module.

In PyF, formatters are in three categories:

Integral numbers, which are numbers without fractional part

Fractional numbers, which are numbers with a fractional part

String, which represents text.

The formatting can be either explicit or implicit. For example:

>>> let x = 10 in [fmt|{x}|]
10

Is an implicit formatting to number, but:

>>> let x = 10 in [fmt|{x:d}|]

Is an explicit formatting to Integral numbers, using d.

Implicit formatting will only format to either Integral, Fractional or text, and this choice is done by the (open) type family PyFCategory.

This modules also provides 3 type class for formatting.

PyfFormatFractional and PyfFormatIntegral are responsible for formatting integral and fractional numbers. Default instances are provided respectively for Real and FormatType. PyFToString is in charge of formatting text.

Default formatting classification

data PyFCategory

The three categories of formatting in PyF

Constructors

PyFIntegral
Format as an integral, no fractional part, precise value

PyFFractional
Format as a fractional, approximate value with a fractional part

PyFString
Format as a string

type family PyFClassify t :: PyFCategory

Classify a type to a PyFCategory This classification will be used to decide which formatting to use when no type specifier in provided.

String formatting

class PyFToString t where

Convert a type to string This is used for the string formatting.

Methods

pyfToString :: t -> String

Real formatting (with optional fractional part)

class PyfFormatFractional a where

Apply a fractional formatting to any type.

A default instance for any Real is provided which internally converts to Double, which may not be efficient or results in rounding errors.

You can provide your own instance and internally use formatFractional which does have the same signatures as pyfFormatFractional but with a RealFrac constraint.

Methods

pyfFormatFractional

Arguments

:: (Integral paddingWidth, Integral precision)

=> Format t t' 'Fractional

-> SignMode
Sign formatting

-> Maybe (paddingWidth, AlignMode k, Char)
Padding

-> Maybe (Int, Char)
Grouping

-> Maybe precision
Precision

-> a

-> String

Integral formatting

class PyfFormatIntegral i where

Apply an integral formatting to any type.

A default instance for any FormatType is provided.

You can provide your own instance and internally use formatIntegral which does have the same signatures as pyfFormatIntegral but with an FormatType constraint.

Methods

pyfFormatIntegral

Arguments

:: Integral paddingWidth

=> Format t t' 'Integral

-> SignMode
Sign formatting

-> Maybe (paddingWidth, AlignMode k, Char)
Padding

-> Maybe (Int, Char)
Grouping

-> i

-> String