PyF.Formatters

Description

Formatters for integral / fractional and strings.

Is support:

For all types:

Grouping of the integral part (i.e: adding a custom char to separate groups of digits)

Padding (left, right, around, and between the sign and the number)

Sign handling (i.e: display the positive sign or not)

For floating:

Precision

Fixed Exponential Generic formatting

For integrals:

Binary Hexa Octal / Character representation

Generic formatting function

formatString

Arguments

:: forall paddingWidth precision. (Integral paddingWidth, Integral precision)

=> Maybe (paddingWidth, AlignMode 'AlignAll, Char)
Padding

-> Maybe precision
Precision (will truncate before padding)

-> String

-> String

Format a string

formatIntegral

Arguments

:: Integral paddingWidth

=> Integral i

=> Format t t' 'Integral

-> SignMode

-> Maybe (paddingWidth, AlignMode k, Char)
Padding

-> Maybe (Int, Char)
Grouping

-> i

-> String

Format an integral number

formatFractional

Arguments

:: (RealFloat f, Integral paddingWidth, Integral precision)

=> Format t t' 'Fractional

-> SignMode

-> Maybe (paddingWidth, AlignMode k, Char)
Padding

-> Maybe (Int, Char)
Grouping

-> Maybe precision
Precision

-> f

-> String

Format a fractional number

Formatter details

data AltStatus

This formatter support alternate version

Constructors

CanAlt

NoAlt

data UpperStatus

This formatter support Upper case version

Constructors

CanUpper

NoUpper

data FormatType

This formatter formats an integral or a fractional

Constructors

Fractional

Integral

data Format (k :: AltStatus) (k' :: UpperStatus) (k'' :: FormatType) where

All the Formatters

Constructors

Decimal :: Format 'NoAlt 'NoUpper 'Integral

Character :: Format 'NoAlt 'NoUpper 'Integral

Binary :: Format 'CanAlt 'NoUpper 'Integral

Hexa :: Format 'CanAlt 'CanUpper 'Integral

Octal :: Format 'CanAlt 'NoUpper 'Integral

Fixed :: Format 'CanAlt 'CanUpper 'Fractional

Exponent :: Format 'CanAlt 'CanUpper 'Fractional

Generic :: Format 'CanAlt 'CanUpper 'Fractional

Percent :: Format 'CanAlt 'NoUpper 'Fractional

Alternate :: Format 'CanAlt u f -> Format 'NoAlt u f

Upper :: Format alt 'CanUpper f -> Format 'NoAlt 'NoUpper f

data SignMode

Sign handling

Constructors

Plus
Display - sign and + sign

Minus
Only display - sign

Space
Display - sign and a space for positive numbers

data AnyAlign where

Existential version of AlignMode

Constructors

AnyAlign :: AlignMode (k :: AlignForString) -> AnyAlign

Internal usage only

data AlignMode (k :: AlignForString) where

Alignement

Constructors

AlignLeft :: AlignMode 'AlignAll
Left padding

AlignRight :: AlignMode 'AlignAll
Right padding

AlignInside :: AlignMode 'AlignNumber
Padding will be added between the sign and the number

AlignCenter :: AlignMode 'AlignAll
Padding will be added around the value

getAlignForString :: AlignMode k -> Maybe (AlignMode 'AlignAll)

data AlignForString

Constructors

AlignAll

AlignNumber