PyF.Internal.QQ

Description

This module uses the python mini language detailed in PythonSyntax to build an template haskell expression representing a formatted string.

Documentation

toExp :: Config -> String -> Q Exp

Parse a string and return a formatter for it

data Config

Configuration for the quasiquoter

Constructors

Config

Fields

delimiters :: Maybe (Char, Char)

What are the delimiters for interpolation. Nothing means no interpolation / formatting.

postProcess :: Q Exp -> Q Exp

Post processing. The input Exp represents a String. Common use case includes using wrapFromString to add fromString in the context of OverloadedStrings.

wrapFromString :: ExpQ -> Q Exp

If OverloadedStrings is enabled, from the input expression with fromString.

expQQ :: String -> (String -> Q Exp) -> QuasiQuoter

Build a quasiquoter for expression